pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.42",
    "streamlit>=1.48.0",
]
